Adjust PokemonList Grid Css
Change Grid css to `auto-fill` the rows instead of `auto-fit`.

Auto-fill: FILLS the row with as many columns as it can fit.
So it creates implicit columns whenever a new column can fit,
because it’s trying to FILL the row with as many columns as it can.
The newly added columns can and may be empty, but they will still
occupy a designated space in the row.

Auto-fit: FITS the CURRENTLY AVAILABLE columns into the space by
expanding them so that they take up any available space.
The browser does that after FILLING that extra space with extra
columns (as with auto-fill ) and then collapsing the empty ones.